refactor: move DuckDuckGoSearchTool tests to a new file and clean up imports
